services/auth-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
prometheus-client==0.21.0
pydantic==2.9.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
